+++ plugin.rb
@@ -2,7 +2,7 @@
 
 # name: discourse-forum-fortress
 # about: Forum Fortress anti-spam protection for Discourse
-# version: 0.1.0-alpha.1
+# version: 0.1.0-alpha.2
 # authors: Forum Fortress
 # url: https://github.com/Forum-Fortress/discourse
 # required_version: 2026.8.0